+++ src/__tests__/tag-test.js
@@ -1,0 +1,9 @@
+let tag = require('../tag')
+
+describe('tag', function() {
+
+  it ('includes the function name', function() {
+    tag(function test() {}).toString().should.include('test')
+  })
+
+})